+++ addons/wms_training/tests/__init__.py
@@ -1,1 +1,1 @@
-from . import test_beginner_mode, test_help_video, test_tour_links
+from . import test_beginner_mode, test_help_product_master, test_help_video, test_tour_links